update: delete button, modal popup lc

+++ database/migrations/2024_01_20_065755_villages.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('villages', function (Blueprint $table){
             $table->id();
-            $table->integer('district_id');
+            $table->bigInteger('district_id');
             $table->string('name');
             // $table->timestamps();
         });
